Delete BaseCCDConfig
