+++ ccd-config-generator/build.gradle
@@ -29,8 +29,6 @@ dependencies {
     testImplementation group: 'org.springframework.boot', name: 'spring-boot-starter-test', version: springBootVersion
 
 
-    testCompile group: 'org.springframework.boot', name: 'spring-boot-starter-mail', version: '2.4.4'
-
     compile group: 'org.apache.commons', name: 'commons-lang3', version: '3.12.0'
     compile group: 'de.cronn', name: 'reflection-util', version: '2.10.0'
     compile group: 'org.objenesis', name: 'objenesis', version: '3.2'
