+++ ccd-config-generator/src/test/java/uk/gov/hmcts/reform/fpl/CCDConfigDevelopment.java
@@ -2,11 +2,7 @@ package uk.gov.hmcts.reform.fpl;
 
 import uk.gov.hmcts.ccd.sdk.api.Webhook;
 
-public class CCDConfigDevelopment extends CCDConfig {
-    @Override
-    public void configure() {
-        super.configure();
-    }
+public class CCDConfigDevelopment extends FPLConfig {
 
     @Override
     protected String webhookConvention(Webhook webhook, String eventId) {
